EPMRPP-97554 || Notifications (#4141)
* EPMRPP-97554 || Notifications

* EPMRPP-97554 || code review fix

* EPMRPP-97554 || lib upgrade

* EPMRPP-97554 || lib upgrade

* EPMRPP-97554 || crf - 2
