+++ app/package.json
@@ -25,7 +25,7 @@
     "@formatjs/intl-pluralrules": "1.3.9",
     "@formatjs/intl-relativetimeformat": "4.5.1",
     "@formatjs/intl-utils": "1.6.0",
-    "@reportportal/ui-kit": "^0.0.1-alpha.26",
+    "@reportportal/ui-kit": "^0.0.1-alpha.38",
     "axios": "1.6.4",
     "c3": "0.7.20",
     "chart.js": "2.9.4",
