+++ app/src/components/main/notification/notificationList/notificationList.scss
@@ -16,29 +16,53 @@
 
 .notification-list {
   position: fixed;
-  bottom: 0;
-  left: 0;
-  right: 0;
+  bottom: 48px;
+  left: 50%;
   opacity: 0.8;
+  max-width: 600px;
+  transform: translateX(-50%);
+}
+
+.notification-item-wrapper {
+  margin-top: 16px;
+  overflow: hidden;
+  opacity: 0;
+  max-height: 0;
+  min-height: 0;
 }
 
 :global {
   .notification-transition-enter {
     opacity: 0.01;
-    height: 0;
+    max-height: 0;
+    min-height: 0;
   }
+
+  .notification-transition-enter-done {
+    opacity: 1;
+    transform: translateY(0);
+    max-height: 80px;
+    min-height: 56px;
+  }
+
   .notification-transition-enter-active {
     opacity: 1;
-    height: 36px;
+    max-height: 80px;
+    min-height: 56px;
     transition: all 200ms ease-in;
   }
+
   .notification-transition-exit {
     opacity: 1;
-    height: 36px;
+    max-height: 80px;
+    min-height: 56px;
   }
+
   .notification-transition-exit-active {
     opacity: 0.01;
-    height: 0;
+    transform: translateY(20px);
     transition: all 700ms ease-in;
+    max-height: 0;
+    min-height: 0;
   }
 }
