--- app/src/components/main/notification/notificationList/notificationListItem/notificationItem.scss
@@ -1,43 +1,0 @@
-@use "sass:color";
-/*!
- * Copyright 2019 EPAM Systems
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-.message-container {
-  overflow: hidden;
-  font-family: $FONT-REGULAR;
-  font-size: 14px;
-  line-height: 16px;
-  text-align: center;
-  color: $COLOR--white;
-  padding: 10px;
-  background: $COLOR--topaz;
-  text-shadow: 0 1px 1px color.adjust($COLOR--topaz, $lightness: -10%);
-  &.success {
-    background: $COLOR--dark-pastel-green;
-    text-shadow: 0 1px 1px color.adjust($COLOR--dark-pastel-green, $lightness: -10%);
-  }
-  &.error {
-    background: $COLOR--tomato;
-    text-shadow: 0 1px 1px color.adjust($COLOR--tomato, $lightness: -10%);
-  }
-  &.opacity-enter-active {
-    opacity: 1;
-    transition: opacity 5500ms ease-in;
-  }
-  &.opacity-leave-active {
-    transition: opacity 5300ms ease-in;
-  }
-}
